gears/info.py: Show Engine rating (size) in its info panel.
It's highly annoying when trying to sort through engines you got
from various salvaged and purchased mecha and trying to find the
largest such engine you can slap into a mecha you are (re)designing.

We potentially can add this info block in the Armor Sensors etc
Component gears as well but those have their size already in the
names of the parts so why bother.