game/content/ghplots/missionbuilder.py: Properly set assistant in BAM_DefeatArmy.
`gears.selector.generate_ace` returns a `gears.base.Mecha`, not a
`gears.base.Character`.

If we load the return value of `generate_ace` directly into the
`"_assistant"` element, then the objective will be named something
like "Defeat Amuro and Buru Buru", naming the mecha of the assistant
rather than the assistant zemself, which seems a bit rude.

The only other use of `generate_ace` in the same file also makes an
effort to extract the pilot and put it in the appropriate element.