gears/__init__.py: Fix disappearing half-cannibalized mechas bug.
Fixes: #126

We used to have a spurious problem where a half-cannibalized mecha
with torso and/or engines missing would *sometimes* disappear from
your party list whenever you moved between scenes.

Turns it only happened sometimes because we were treating it as a
combat-destroyed mecha, and the repair roll would kick in, sometimes
saving the mecha, sometimes failing the repair roll and losing the
mecha.

Fix by only destroying mecha on the map.