gears/attackattributes.py: Rescale ChargeAttack for the new damage scaling.
Looks like the main reason why Charge *still* seems overpowered is because
when we rescaled damage vs. health, we forgot to rescale Charge as well.
We only rescaled `get_basic_attack` for all attacking gears, but
`attackattributes.ChargeAttack` overrides the raw value instead of basing
on the raw value.

Prior to this change, Charge meant more than 2.5x the normal damage, a
massive bonus compared to the +2/3 damage that Charge did prior to the
rescale.

With this change, Charge adds +75% more damage, and should adapt to future
damage rescalings as well, targeting +2/3 damage.